requirements.txt
certifi==2019.3.9
Django==2.2
django-debug-toolbar==1.11
django-extensions==2.1.6
djangorestframework==3.9.2
factory-boy==2.11.1
Faker==1.0.5
pipenv==2018.11.26
psycopg2==2.8.2
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
virtualenv==16.5.0
virtualenv-clone==0.5.3
